IR2: Move all types into 1 module
