--- lib/Besra/Types/IR2/Number.hs
@@ -1,8 +1,0 @@
-
-module Besra.Types.IR2.Number ( Number(..) ) where
-
-import Protolude
-
-
-newtype Number = Number Int
-  deriving (Eq, Show)
