--- lib/Besra/Types/IR2/String.hs
@@ -1,9 +1,0 @@
-
-module Besra.Types.IR2.String ( String(..) ) where
-
-import Protolude
-
-
-newtype String = String Text
-  deriving (Eq, Show)
-
